\name{SimulateHiGeoSSE}
\alias{SimulateHiGeoSSE}
\title{Simulate under a HiGeoSSE model}
\description{Simulation function for the HiGeoSSE model}
\usage{
SimulateHiGeoSSE(pars, hidden.areas=1, x0="AB0", max.taxa=Inf, max.time=Inf, include.extinct=FALSE, return.HiGeoSSE_pars=FALSE, override.safeties=FALSE)}       
\arguments{
\item{pars}{an object of class "HiGeoSSE_pars" generated by
  \code{SimulateHiGeoSSE}. See 'Details'.}
\item{hidden.ares}{number of hidden areas in the model. Default is 1.}
\item{x0}{a character value for the starting state. Default is "AB0".}
\item{max.taxa}{have the simulation stop at \code{max.taxa} surviving taxa.}
\item{max.time}{have the simulation stop at \code{max.time} height
  (including stem).} 
\item{include.extinct}{whether the simulation should keep extinct
  lineages. Default is FALSE.}
\item{return.HiGeoSSE_pars}{whether to (only) return a "hiGeoSSE_pars" object
  in the correct format to use in this function. See 'Details'.}
\item{override.safeties}{simulate even if there is no limit on number of
taxa or depth of tree.}
}
\details{
This function uses \code{diversitree::tree.classe} function to simulate
the tree under the GeoSSE model with hidden areas. The parameter
\code{pars} is a list of class "HiGeoSSE_pars" that can be generated
using this same function. The best way is to first run
\code{SimulateHiGeoSSE} with empty parameters and
\code{return.HiGeoSSE_pars=TRUE} and with \code{hidden.areas} equal to
the number of desired hidden areas in the model. This will return an
object in list format with a set of four matrices. Each matrix controls
a different set of parameters of the model. All parameters are set to 0
and parameters not used by the model are set to NA. To prepare the
simulation, edit these matrices and change the desired parameter
values. Leaving a parameter equal to 0 is the same as reducing the model
by eliminating this parameter of the model.
}
\value{
	This returns a list, with the following elements:
	\item{$phy}{a phylogeny of class phylo.}
	\item{$data}{a named vector with the tip states.}
	\item{$sim.attempts}{number of times that the simulation tried
	  before reaching desired termination condition.}
      }      
\examples{
\dontrun{
## Get the a list with the correct parameters to run the simulation:
pars <- SimulateHiGeoSSE(hidden.areas=1, return.HiGeoSSE_pars=TRUE)
## Edit the parameter values:
pars$model.pars[1:3,] <- 0.1
pars$model.pars[4:5,] <- 0.05
pars$model.pars[6:7,] <- 0.01
pars$q.AB[1,2] <- pars$q.AB[2,1] <- 0.01
pars$q.A[1,2] <- pars$q.A[2,1] <- 0.02
pars$q.B[1,2] <- pars$q.B[2,1] <- 0.02
## Run the simulation:
sim <- SimulateHiGeoSSE(pars=pars, hidden.areas=1, x0="AB0", max.taxa=100)
}
}
\references{}
\author{Daniel Caetano}
\keyword{simulation, GeoSSE, HiGeoSSE}
